+++ frontend/public/js/newFile.js
@@ -1,4 +1,4 @@
-$("#new-button").on("click", async () => {
+$("#new-file").on("click", async () => {
   var response = await fetch("/new", { method: "POST" });
   console.log(response);
   if (response.url) window.location = response.url;