+++ frontend/public/js/uploadButton.js
@@ -1,5 +1,0 @@
-$("#upload-button").on("click", () => $("#input").trigger("click"));
-
-$("#input").on("input", () => {
-  $("#input-form").submit();
-});