+++ frontend/public/js/uploadFile.js
@@ -1,0 +1,13 @@
+$("#upload-file").on("click", () => {
+  $("#input").attr({
+    type: "file",
+    name: "file",
+    accept: "text/*",
+  });
+
+  $("#input")
+    .trigger("click")
+    .on("input", () => {
+      $("#input-form").submit();
+    });
+});